feat: separating the context from the component
We've basically separated out all the code responsible for taking the transactions from the context and then iterating over this data using the reduce function. All this code has gone into a hook file, so inside the Summary component we only need to execute the useSummary function to have access to the summary.